python-service/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pydantic==2.9.0
akshare>=1.16.72
yfinance>=0.2.44
pandas>=2.2.0
numpy>=1.26.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.3.0
crewai>=0.100.0
crewai-tools>=0.33.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
openai>=1.55.0
python-dotenv>=1.0.1
pymongo>=4.8.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=6.0.0
